drift-guard-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drift-guard-agent"
version = "0.1.0"
description = "LangGraph-powered AI agent for drift-guard: discovers impacted consumers, explains breaking changes, and notifies teams."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-anthropic>=0.3",
    "anthropic>=0.50",
    "httpx>=0.27",
    "gitpython>=3.1",
    "click>=8.1",
]

[project.scripts]
drift-guard-agent = "drift_guard_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["drift_guard_agent"]
